items/serializers.py
from rest_framework import serializers

from .models import Category, Item
# view = object(request) - response

class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = "__all__"

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = "__all__"