requirements.txt
django
djangorestframework
Pillow
django-cors-headers
drf-yasg